refactor: inline serve helpers into handler
