+++ packages/cli/src/handlers/serve.test.ts
@@ -1,19 +1,22 @@
-import { strictEqual } from "node:assert";
+import { rejects, strictEqual } from "node:assert";
+import { mkdir, mkdtemp, rm, writeFile } from "node:fs/promises";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
 import { afterEach, describe, it, vi } from "vitest";
 
 const consoleLogMock = vi.fn();
 const consoleErrorMock = vi.fn();
-const resolveServeServerEntryMock = vi.fn();
-const startServeServerMock = vi.fn();
 const exitWithErrorMock = vi.fn((message: string) => {
   consoleErrorMock(`Error: ${message}`);
   throw new Error(`Exit: ${message}`);
 });
 
+const temporaryDirectories: string[] = [];
 const originalHost = process.env.HOST;
 const originalNitroHost = process.env.NITRO_HOST;
 const originalPort = process.env.PORT;
 const originalNitroPort = process.env.NITRO_PORT;
+const originalArgv = [...process.argv];
 
 vi.doMock("../output.ts", () => ({
   output: {
@@ -26,15 +29,11 @@ vi.doMock("../errors.ts", () => ({
   exitWithError: exitWithErrorMock,
 }));
 
-vi.doMock("../serve.ts", () => ({
-  resolveServeServerEntry: resolveServeServerEntryMock,
-  startServeServer: startServeServerMock,
-}));
+const { resolveServeServerEntry, serveHandler } = await import("./serve.ts");
 
-const { serveHandler } = await import("./serve.ts");
-
-afterEach(() => {
+afterEach(async () => {
   vi.clearAllMocks();
+  process.argv = [...originalArgv];
 
   if (originalHost === undefined) {
     delete process.env.HOST;
@@ -59,26 +58,130 @@ afterEach(() => {
   } else {
     process.env.NITRO_PORT = originalNitroPort;
   }
+
+  await Promise.all(
+    temporaryDirectories
+      .splice(0)
+      .map((directory) => rm(directory, { recursive: true, force: true })),
+  );
+});
+
+async function createTemporaryDirectory(): Promise<string> {
+  const directory = await mkdtemp(join(tmpdir(), "phantom-serve-handler-"));
+  temporaryDirectories.push(directory);
+  return directory;
+}
+
+async function createBundledCliFixture(): Promise<{
+  cliEntry: string;
+  serverEntry: string;
+}> {
+  const directory = await createTemporaryDirectory();
+  const cliEntry = join(directory, "packages", "cli", "dist", "phantom.js");
+  const serverEntry = join(
+    directory,
+    "packages",
+    "cli",
+    "dist",
+    "app",
+    ".output",
+    "server",
+    "index.mjs",
+  );
+
+  await mkdir(
+    join(directory, "packages", "cli", "dist", "app", ".output", "server"),
+    {
+      recursive: true,
+    },
+  );
+  await writeFile(cliEntry, "");
+  await writeFile(serverEntry, "");
+
+  return { cliEntry, serverEntry };
+}
+
+describe("resolveServeServerEntry", () => {
+  it("finds the app server entry from the bundled CLI entrypoint", async () => {
+    const { cliEntry, serverEntry } = await createBundledCliFixture();
+
+    strictEqual(await resolveServeServerEntry(cliEntry), serverEntry);
+  });
+
+  it("throws when the bundled server entry is missing", async () => {
+    const directory = await createTemporaryDirectory();
+    const cliEntry = join(directory, "packages", "cli", "dist", "phantom.js");
+
+    await mkdir(join(directory, "packages", "cli", "dist"), {
+      recursive: true,
+    });
+    await writeFile(cliEntry, "");
+
+    await rejects(
+      resolveServeServerEntry(cliEntry),
+      /Could not find Phantom server assets/,
+    );
+  });
+
+  it("throws when called from the source CLI entrypoint", async () => {
+    const directory = await createTemporaryDirectory();
+    const cliEntry = join(
+      directory,
+      "packages",
+      "cli",
+      "src",
+      "bin",
+      "phantom.ts",
+    );
+    const serverEntry = join(
+      directory,
+      "packages",
+      "cli",
+      "dist",
+      "app",
+      ".output",
+      "server",
+      "index.mjs",
+    );
+
+    await mkdir(join(directory, "packages", "cli", "src", "bin"), {
+      recursive: true,
+    });
+    await mkdir(
+      join(directory, "packages", "cli", "dist", "app", ".output", "server"),
+      {
+        recursive: true,
+      },
+    );
+    await writeFile(cliEntry, "");
+    await writeFile(serverEntry, "");
+
+    await rejects(
+      resolveServeServerEntry(cliEntry),
+      /Could not find Phantom server assets/,
+    );
+  });
 });
 
 describe("serveHandler", () => {
   it("uses port 9640 by default", async () => {
-    resolveServeServerEntryMock.mockResolvedValue("/bundle/server/index.mjs");
+    const { cliEntry, serverEntry } = await createBundledCliFixture();
+    process.argv[1] = cliEntry;
 
     await serveHandler([]);
 
     strictEqual(process.env.PORT, "9640");
     strictEqual(process.env.NITRO_PORT, "9640");
-    strictEqual(resolveServeServerEntryMock.mock.calls.length, 1);
-    strictEqual(startServeServerMock.mock.calls.length, 1);
+    strictEqual(consoleLogMock.mock.calls.length, 1);
     strictEqual(
-      startServeServerMock.mock.calls[0][0],
-      "/bundle/server/index.mjs",
+      consoleLogMock.mock.calls[0][0],
+      `Starting Phantom server from ${serverEntry}`,
     );
   });
 
   it("lets --port override the default port", async () => {
-    resolveServeServerEntryMock.mockResolvedValue("/bundle/server/index.mjs");
+    const { cliEntry } = await createBundledCliFixture();
+    process.argv[1] = cliEntry;
 
     await serveHandler(["--port", "4100"]);
